@@ -5,7 +5,7 @@ apply plugin: 'kotlin-android-extensions'
 project.ext {
     supportLibraryVersion = "28.0.0"
     androidArchVersion = "1.0.0-rc1"
-    daggerVersion = "2.22"
+    daggerVersion = "2.35.1"
     rxJavaVersion = "2.1.0"
     rxAndroidVersion = "2.0.1"
     priorityJobQueueVersion = "2.0.1"
@@ -25,11 +25,11 @@ android {
             storePassword 'odedtech@12'
         }
     }
-    compileSdkVersion 29
+    compileSdkVersion 30
     defaultConfig {
         applicationId "com.odedtech.mff"
         minSdkVersion 21
-        targetSdkVersion 29
+        targetSdkVersion 30
         versionCode 2
         versionName "1.1"
         multiDexEnabled true
@@ -108,17 +108,18 @@ dependencies {
     //noinspection GradleCompatible
     implementation 'com.android.support:support-v4:28.0.0'
     implementation project(path: ':shuftipro-release')
+    implementation 'com.google.firebase:protolite-well-known-types:18.0.0'
     testImplementation 'junit:junit:4.12'
     //noinspection GradleCompatible
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
 
     // butter knife
-    implementation 'com.jakewharton:butterknife:8.8.1'
+    implementation 'com.jakewharton:butterknife:10.1.0'
     implementation 'com.google.android.gms:play-services-maps:16.1.0'
     implementation 'com.google.android.gms:play-services-location:16.0.0'
     implementation 'com.google.android.gms:play-services-auth:16.0.0'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.1.0'
 
     // circular image view
     implementation 'de.hdodenhof:circleimageview:2.2.0'
@@ -211,7 +212,7 @@ dependencies {
     implementation files('libs/ProwessPride_V1.12.jar')
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
-    implementation 'com.facebook.android:facebook-login:[5,6)'
+   // implementation 'com.facebook.android:facebook-login:[5,6)'
     implementation 'com.github.mukeshsolanki:android-otpview-pinview:2.0.3'
 
     //implementation 'com.github.lecho:hellocharts-android:v1.5.8@arr'
@@ -219,5 +220,9 @@ dependencies {
     implementation 'com.github.gcacace:signature-pad:1.3.1'
 
     implementation 'com.squareup.picasso:picasso:2.71828'
+
+    //Pdf Reader
+    implementation 'com.github.barteksc:android-pdf-viewer:2.8.2'
+    implementation 'com.mindorks.android:prdownloader:0.6.0'
 
 }
