@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript { 
-    ext.kotlin_version = '1.3.41'
+    ext.kotlin_version = '1.6.21'
 
     
     repositories {
