@@ -1,23 +1,48 @@
 package vas;
 
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.widget.AppCompatButton;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.odedtech.mff.mffapp.R;
 
+import Utilities.AlertDialogUtils;
+import Utilities.PreferenceConnector;
 import base.BaseFragment;
+import login.LoginActivity;
 
 public class VasFragment extends BaseFragment {
+    public AppCompatButton logoutButton;
+    private View view;
+
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_vas_new, container, false);
+        view = inflater.inflate(R.layout.fragment_vas_new, container, false);
+        logoutButton = view.findViewById(R.id.logout_button);
+
+        logoutButton.setOnClickListener(v -> {
+            AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getContext());
+            alertDialogBuilder.setMessage("Are you sure, You want to logout?");
+            alertDialogBuilder.setPositiveButton("yes", (dialog, which) -> {
+                callLogout();
+            });
+
+            alertDialogBuilder.setNegativeButton("No", (dialog, which) -> {
+            });
+            AlertDialog alertDialog = alertDialogBuilder.create();
+            alertDialog.show();
+        });
         return view;
     }
 
@@ -31,5 +56,13 @@ public class VasFragment extends BaseFragment {
         view.findViewById(R.id.card_block).setOnClickListener(view1 -> {
             startActivity(new Intent(getActivity(), BlockCardActivity.class));
         });
+    }
+
+    private void callLogout() {
+        PreferenceConnector.clearPref(getContext());
+        Intent intent = new Intent(view.getContext(), LoginActivity.class);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+        startActivity(intent);
+        requireActivity().finish();
     }
 }
