@@ -2,17 +2,18 @@ package base;
 
 import android.app.ProgressDialog;
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-
 import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.google.gson.Gson;
 import com.odedtech.mff.mffapp.R;
+
+import java.io.IOException;
 
 import Utilities.PreferenceConnector;
 import login.LoginActivity;
 import retrofit2.Response;
-
-import java.io.IOException;
 
 public class BaseActivity extends AppCompatActivity {
 
@@ -23,14 +24,16 @@ public class BaseActivity extends AppCompatActivity {
         progressDialog.setMessage("Loading...");
         progressDialog.setCancelable(false);
         progressDialog.setCanceledOnTouchOutside(false);
-        if(progressDialog == null) {
+        if (progressDialog == null) {
+            assert false;
             progressDialog.show();
         }
     }
 
     public void dismissLoading() {
-        if (progressDialog != null && progressDialog.isShowing()) {
-            progressDialog.dismiss();
+        if (progressDialog != null) {
+            if (progressDialog.isShowing())
+                progressDialog.dismiss();
         }
     }
 
@@ -42,8 +45,9 @@ public class BaseActivity extends AppCompatActivity {
     }
 
     protected void checkResponseError(Response response) {
-        if (response.code() == 401) {
+        if (response.code() == 401 && response.body() != null) {
             try {
+                assert response.errorBody() != null;
                 MFFErrorResponse errorResponse = new Gson()
                         .fromJson(response.errorBody().string(), MFFErrorResponse.class);
                 if (errorResponse.error != null && errorResponse.error.equals("invalid_token")) {
