@@ -2,9 +2,6 @@ package base;
 
 import android.app.DatePickerDialog;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,11 +9,16 @@ import android.view.ViewGroup;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
 import com.google.gson.Gson;
 import com.odedtech.mff.mffapp.R;
 
 import java.io.IOException;
 import java.util.Calendar;
+import java.util.Objects;
 
 import Utilities.UtilityMethods;
 import retrofit2.Response;
@@ -29,11 +31,11 @@ public class BaseFragment extends Fragment {
     }
 
     public void showLoading() {
-        ((BaseActivity) getActivity()).showLoading();
+        ((BaseActivity) requireActivity()).showLoading();
     }
 
     public void dismissLoading() {
-        ((BaseActivity) getActivity()).dismissLoading();
+        ((BaseActivity) requireActivity()).dismissLoading();
     }
 
     protected void showDatePickerDialog(EditText editText) {
