@@ -2,8 +2,8 @@ package client.login
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.view.View
+import androidx.appcompat.app.AppCompatActivity
 import com.odedtech.mff.mffapp.R
 
 class ClientOTPActivity : AppCompatActivity() {
