@@ -2,10 +2,11 @@ package com.odedtech.mff.mff_android;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 import com.odedtech.mff.mffapp.R;
 
