@@ -1,9 +1,9 @@
 package data;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
 import android.content.Context;
+
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
 
 import Utilities.Constants;
 import loans.model.LoanCollection;
