@@ -1,21 +1,18 @@
 package loans;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.CheckBox;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.odedtech.mff.mffapp.R;
 
 import java.util.List;
 
-import loans.model.Datum;
 import loans.model.SearchData;
-import loans.model.SearchResponse;
-import vas.AccountsBlockAdapter;
 
 public class SearchAdapter extends RecyclerView.Adapter<SearchAdapter.ViewModel> {
 
