@@ -1,15 +1,24 @@
 package onboard;
 
 import Utilities.PreferenceConnector;
+
+import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.bumptech.glide.Glide;
 import com.odedtech.mff.mffapp.R;
+
+import org.jetbrains.annotations.NotNull;
+
 import de.hdodenhof.circleimageview.CircleImageView;
 
 import java.util.ArrayList;
@@ -46,7 +55,7 @@ public class ProfilesListAdapter extends RecyclerView.Adapter<ProfilesListAdapte
     }
 
     @Override
-    public void onBindViewHolder(ViewModel holder, int position) {
+    public void onBindViewHolder(@NonNull @NotNull ViewModel holder, @SuppressLint("RecyclerView") int position) {
         ClientDataDTO clientDataDTO = filteredList.get(position);
         holder.nameTV.setText(clientDataDTO.name);
         holder.textRole.setText(clientDataDTO.identifier);
