@@ -2,12 +2,15 @@ package onboard;
 
 import Utilities.PreferenceConnector;
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.bumptech.glide.Glide;
 import com.odedtech.mff.mffapp.R;
 import de.hdodenhof.circleimageview.CircleImageView;
